syn sync maxlines
This somehow improves accuracy of syntax highlighting in very large
files that contain many `markdownCodeBlock`s.

Previously, when editing the same buffer in separate windows, an edit at
the upper part of the buffer in one window frequently broke the
highlighting (wrongly highlight to `markdownCodeBlock`) in another
window that shows lower part of the buffer.